Addressing review comments and created separate function in pgsql
